refactor(core): split responsibilities in index.ts and improve app version handling
- Moved health check and shutdown-related logic into separate files
- Improved clarity and structure of app version retrieval
- Simplified and cleaned up index.ts for better separation of concerns
