+++ .vscode/settings.json
@@ -18,7 +18,7 @@
     "**/pnpm-lock.yaml": true,
     "**/LICENSE": true,
     "**/nginx": true,
-    "**/nodemon.json": true,
+    // "**/nodemon.json": true,
     "**/README.md": true
   },
