+++ apps/gateway/src/utils/server/getAppVersion.ts
@@ -1,0 +1,5 @@
+import pkg from '../../../package.json' with { type: 'json' };
+
+export function getAppVersion(): string {
+  return pkg.version || 'unknown';
+}
