+++ apps/gateway/src/utils/server/index.ts
@@ -1,0 +1,3 @@
+export * from './getAppVersion.js';
+export * from './health.js';
+export * from './shutdown.js';
