+++ apps/gateway/src/utils/server/shutdown.ts
@@ -1,0 +1,9 @@
+let shuttingDown = false;
+
+export function markShuttingDown(): void {
+  shuttingDown = true;
+}
+
+export function isShuttingDown(): boolean {
+  return shuttingDown;
+}
