+++ services/auth-service/src/app.ts
@@ -10,14 +10,12 @@ const app = express();
 // wyłączenie X-Powered-By
 app.disable('x-powered-by');
 app.use(express.json()); // parsowanie JSON w body
-if (process.env.NODE_ENV === 'development') {
-  app.use(requestLoggerDev);
-}
-
-app.use((req, res, next) => {
-  logger.info(`HTTP ${req.method} ${req.url} - IP: ${req.ip}`);
-  next();
-});
+app.use(
+  requestLoggerDev({
+    logger,
+    isDev: env.NODE_ENV === 'development',
+  })
+);
  
 // Prosty endpoint, który zwraca losową odpowiedź
 app.post('/check-email', (req, res) => {
